// src/modules/internalizacion/interfaces/UpdateInternalizacion.ts
export interface InternalizacionUpdate{
    SED_ID: number;
    ICAR_ID: number;
    INTER_CONV_CODIGO: string;
    INTER_CONV_NOMBRE: string;
    INTER_CONTRAPARTE: string;
    PAI_ID: number;
    PTIP_ID: number;
    INTER_PAPELLIDO: string;
    INTER_PNOMBRE: string;
    INTER_ACTIVIDADES: string;
    INTER_FINICIO: string;
    INTER_FFIN: string;
    ARE_ID: number;
    FIN_ID: number;
    INTER_ADOCENCIA: string;
    INTER_AINVESTIGACION: string;
    INTER_AVINCULACION: string;
    INTER_AADMINISTRATIVO: string;
    INTER_AINTERNACIONALIZACION: string;
    INTER_INDICADOR_1: string;
    INTER_INDICADOR_2: string;
    INTER_INDICADOR_3: string;
    INTER_INDICADOR_4: string;
    INTER_INDICADOR_5: string;
    INTER_INDICADOR_6: string;
    INTER_INDICADOR_7: string;
    INTER_INDICADOR_8: string;
    INTER_INDICADOR_9: string;
    INTER_EVIDENCIA: string;
    INTER_ESTADO: number;
}